requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
flask==3.0.0
flask-cors==4.0.0
